Consistent generation of Quagga router ids
The router id is stored and computed in Zebra Daemon.This is the
only daemon that is guaranteed to be present if any of ospfd, ospf6d
and bgpd is present. The default choice for the router id is still
to take the most-visible IPv4 address as router id. In the case of
an IPv6-only network, an incrementing router id is assigned.
However, this is still not sufficient if some router ids are
explicitly set in one of daemons. Router ids should not collide with
these "explicit" router ids. In order to do so, a check is performed
before assigning a router id.