Add IPv6 support in the CLI (#2)
* Enhancing ping-related commands from Mininet CLI

These commands will also check IPv6 connectivity if IPv6 is enabled and
if IPv4 is not enabled, it does not check IPv4 connectivity.

* Enhance Mininet CLI auto-replacement by addresses

Node names in commands can be replaced by their IPv6 address if any
if IPv4 is not enabled or if no IPv4 addresses was available for
this node.

* Change IP address replacement strategy

We select only one IP version for these automatic replacements. The IP
version chosen is first restricted by the addresses available on the
first node. Then, we choose the IP version that enables as many
replacements as possible. We use IPv4 as a tie-break.

* Improve auto-replacement of node names

Error is raised if one of the nodes is not compatible to the chosen IP
version.

* Improves ping strategy based on the IP replacement strategy

* Ping improvements

-> Add v4-only and v6-only pings
-> fix the ping function handling of incompatible hosts
-> Imrpove doc of the ping function

* Handles ping6 deprecation + improves doc ping