Enable to change the prefix length of an address (#5)
It cleans the old addresses before adding the new ones. This allows an
existing address to be passed as argument to _set_ip(). For instance,
we may want to change its prefix length.
Before this commit, the 'ip address add' command would fail
since the address already exists and then, the address would be
removed at the cleaning step.