Support facebooks OpenR routing plattform (#18)
* [OpenR] Add basic configuration to support OpenR

https://code.fb.com/connectivity/introducing-open-r-a-new-modular-routing-platform/

This commits contains all changes required to run an OpenR router in
ipmininet. The openr openrd config files have been copied from the ospf
daemons and work analogously.

* [OpenR] Add network example and docu

Add a small OpenR network topology and document it in an appropriate
section in the README.

* [OpenR] Add basic configuration to support OpenR

https://code.fb.com/connectivity/introducing-open-r-a-new-modular-routing-platform/

This commits contains all changes required to run an OpenR router in
ipmininet. The openr openrd config files have been copied from the ospf
daemons and work analogously.

* [OpenR] Add network example and docu

Add a small OpenR network topology and document it in an appropriate
section in the README.

* [OpenR] Document daemon parameters

https://github.com/cnp3/ipmininet/pull/18#discussion_r274382677

Document the parameters and default values passed to the OpenR daemon
with pydoc.