[installer] Add install option for OpenR
Enhance the install script to build and install OpenR. To run OpenR in
ipmininet "routers" there are two main requirements for the host system:

1. OpenR and its dependencies are installed
2. IPv6 is enabled (disabled by the mininet install script)

The OpenR build requires several steps:

1. Clone OpenR
2. Generate the build script from python scripts
3. Make the generated build script executable
4. Build and install OpenR using the build script

IPv6 support is re-enabled by changing/ removing the kernel parameters
in `/etc/default/grub` and `/etc/sysctl.conf`.

Limitations: Currently the OpenR installation is only supported on
Ubuntu. Moreover, the build takes a substantial amount of time and
memory. Therefore we decided to make the OpenR installation optional
(will not be installed with '-a')