Fix find_peer to find the closest interface
The dfs used previously caused iBGP peering failure on topologies with
non-unit IGP costs.

The address configured for a peer was not the source address used
by this peer to contact the BGP daemon. Therefore, the daemon rejected
the connection attempt.

This fixes the issue by taking the IGP cost into account to identify the
closest interface. Note that it might not be sufficient if the routing
inside the AS is static.