[mininet vm] Checkout mininet with release tag
The mininet release of the vm build script is pinned to a certain tag.
At the moment, this tag is only used for the download of the
'install-mininet-vm.sh' script. To chekcout mininet with the same tag,
the option has to be passed to the script.
See:
https://github.com/mininet/mininet/blob/master/util/vm/install-mininet-vm.sh#L40

Pass the release tag to the mininet vm script.